requirements.txt
flask
pandas 
pymongo[srv]
pymongo
python-dotenv
